+++ how-to/customize-home-templates/client/src/integrations/emoji/templates.ts
@@ -1,22 +1,22 @@
 import { ButtonStyle, TemplateFragment } from "@openfin/workspace";
 import { createButton, createContainer, createText } from "../../templates";
-import { getCurrentTheme } from "../../themes";
+import { getCurrentPalette } from "../../themes";
 
 export async function getEmojiTemplate(actions: {
 	copyEmojiAction: string;
 	copyKeyAction: string;
 	detailsAction: string;
 }): Promise<TemplateFragment> {
-	const theme = await getCurrentTheme();
+	const palette = await getCurrentPalette();
 
 	return createContainer(
 		"column",
 		[
-			await createText("keyTitle", 12, { color: theme.palette.brandPrimary, fontWeight: "bold" }),
+			await createText("keyTitle", 12, { color: palette.brandPrimary, fontWeight: "bold" }),
 			await createContainer(
 				"row",
 				[
-					await createText("key", 12, { color: theme.palette.textDefault, wordBreak: "break-all" }),
+					await createText("key", 12, { color: palette.textDefault, wordBreak: "break-all" }),
 					await createButton(ButtonStyle.Secondary, "copyKeyTitle", actions.copyKeyAction, {
 						fontSize: "12px"
 					})
@@ -29,11 +29,11 @@ export async function getEmojiTemplate(actions: {
 				}
 			),
 
-			await createText("emojiTitle", 12, { color: theme.palette.brandPrimary, fontWeight: "bold" }),
+			await createText("emojiTitle", 12, { color: palette.brandPrimary, fontWeight: "bold" }),
 			await createContainer(
 				"row",
 				[
-					await createText("emoji", 32, { color: theme.palette.textDefault }),
+					await createText("emoji", 32, { color: palette.textDefault }),
 					await createButton(ButtonStyle.Secondary, "copyEmojiTitle", actions.copyEmojiAction, {
 						fontSize: "12px"
 					})